+++ packages/forward-builders/tsup.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "tsup";
 import { fixImportsPlugin } from "esbuild-fix-imports-plugin";
 
 export default defineConfig({
-  entry: ["./src/index.ts"],
+  entry: ["./src/index.ts", "./src/config.ts"],
   format: ["esm"],
   target: "es2020",
   outDir: "./dist/packages/forward-builders/src",
@@ -11,5 +11,10 @@ export default defineConfig({
   sourcemap: true,
   esbuildPlugins: [fixImportsPlugin()],
   platform: "node",
-  external: ["@meteora-ag/dlmm", "@raydium-io/raydium-sdk-v2"],
+  external: [
+    "@meteora-ag/dlmm",
+    "@raydium-io/raydium-sdk-v2",
+    "@orca-so/whirlpools",
+    "@solana/kit",
+  ],
 });